Collect processor info with and without smbios usage

+++ client/hwlib/src/models/devices.rs
@@ -80,7 +80,7 @@ pub struct PCIPeripheral {
 #[derive(Serialize, Deserialize, Debug)]
 pub struct Processor {
     pub codename: String,
-    pub frequency: u16,
+    pub frequency: u64,
     pub manufacturer: String,
     pub version: String,
 }
